Вынос конфигурации в отдельные файлы

+++ src/index.ts
@@ -1,7 +1,6 @@
 import Knex from 'knex'
 import { knexSnakeCaseMappers, Model } from 'objection'
-import * as path from 'path'
-import Telegraf, { BotConfig } from 'telegraf'
+import Telegraf from 'telegraf'
 import * as Sentry from '@sentry/node'
 import { captureException, configureScope } from '@sentry/node'
 import * as url from 'url'
@@ -16,15 +15,16 @@ import bindConfig from './helpers/bindConfig'
 import { setupReferralMiddleware } from './middlewares/referralMiddleware'
 import sentryExtraFromCtx from './helpers/sentryExtraFromCtx'
 import User from './models/User'
+import { BotConfig } from 'bot-config'
 
-const config: BotConfig = require(path.resolve(__dirname, '../config/general.json'))
+const config = require('config') as BotConfig
 const debug = require('debug')('bot')
 
-Sentry.init({ dsn: process.env.SENTRY_DSN })
+Sentry.init({ dsn: config.sentry.dsn })
 
 async function setupDb() {
     const knex = Knex({
-        ...require('../knexfile.js')[process.env.NODE_ENV || 'development'],
+        ...config.database,
         ...knexSnakeCaseMappers()
     })
     Model.knex(knex)
@@ -34,7 +34,7 @@ async function setupDb() {
 }
 
 async function setupBot() {
-    const bot = new Telegraf(process.env.TG_BOT_TOKEN)
+    const bot = new Telegraf(config.telegram.token)
 
     bot.use(async (ctx, next) => {
         configureScope((scope) => {
@@ -56,18 +56,14 @@ async function setupBot() {
     // TODO удалить потом эту команду
     bot.command('skidoo', async (ctx) => {
         await ctx.reply(`Вы изгнаны из команды ${ctx.user?.team.name}`)
-        if (ctx.user)
-            await User.query().deleteById(ctx.user.$id())
+        if (ctx.user) await User.query().deleteById(ctx.user.$id())
     })
-
-
-    const gdriveServiceAccount = JSON.parse(process.env.GDRIVE_SERVICE_ACCOUNT)
 
     bindConfig(bot, config)
 
-    const redis = bindSession(bot, process.env.REDIS_URL)
-    const gdrive = await bindGDrive(bot, gdriveServiceAccount, config.gdrive.rootDirId)
-    const trello = await bindTrello(bot, process.env.TRELLO_TOKEN_SECRET)
+    const redis = bindSession(bot, config.redis)
+    const gdrive = await bindGDrive(bot, config.gdrive.serviceAccount, config.gdrive.rootDirId)
+    const trello = await bindTrello(bot, config.trello['appKey:token'])
 
     bot.use(sentryExtraFromCtx('session'))
 
@@ -78,14 +74,14 @@ async function setupBot() {
     // Теперь так делать нельзя
     bot.telegram.webhookReply = false
 
-    if (!process.env.WEBHOOK_URL) {
+    if (!config.telegram.webhook) {
         await bot.telegram.deleteWebhook()
         bot.startPolling()
         debug('Bot started with Longpolling')
     } else {
-        const webhook = new url.URL(process.env.WEBHOOK_URL)
+        const webhook = new url.URL(config.telegram.webhook)
         await bot.telegram.setWebhook(webhook.href)
-        bot.startWebhook(webhook.pathname, null, +process.env.PORT)
+        bot.startWebhook(webhook.pathname, null, +config.server.port)
         debug('Bot started with WebHook on ' + webhook.href)
     }
